requirements.txt
discord.py>=2.3.2
docker>=6.0.0
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0